requirements.txt
boto3
celery[redis]
channels
channels-redis
Django
django-allauth
django-colorfield
django-cors-headers
django-dbbackup
django-filter
django-request
django-storages
djangorestframework
djangorestframework-simplejwt
dj-rest-auth[with_social]
gunicorn
pillow
psycopg2-binary
python-decouple
stripe
uvicorn[standard]

# * Without this package websockets cannot be used in production
# *
# *  - WARNING:  Unsupported upgrade request.
# *  - WARNING:  No supported WebSocket library detected. Please use 'pip install uvicorn[standard]', or install 'websockets' or 'wsproto' manually.
# *
# * Probably uvicorn is installed instead of uvicorn[standard] so we have
# * to manually add dependency 'websockets'.
# *
# * If similar error occurs, try adding dependency 'uvloop'
# * For some reason uvicorn[standard] is not installed, but only uvicorn, so we add this package
websockets

# ! google-api-python-client is used for FIX - remove when resolved (see https://gonzafirewall.medium.com/google-oauth2-and-django-rest-auth-92b0d8f70575)
google-api-python-client
